Add files via upload
The files originated from: https://github.com/voloviq/meta-licheepizero.git

This includes the meta-licheepizero that I used throughout my internship. It includes kernel patches for extended USB connections. The code might be messy and wrong. Keep in mind that this is an old coding and I stopped working on this in 2022.